Combine Z3 and local solution counting
Having to write constraints in two "languages" (Z3 and own exppression
system) in parallel is annoying and opens the door to inconsistencies.
Thus, we create sub-classes of our expressions which also contain the
matching Z3 expression. The same goes for the "problem", i.e., the central
class storing the constraints and allowing solving/optimization.