Predict density and density evolution for sampled_voxel data
This dataset has 300 time steps and thus can make predictions over larger
time spans, which increases prediction quality. Th code is very similar
to the notebook for the other dataset. The prediction pipeline actually is
the same, but we need to consider the different feature set/feature names
and of course our flexible delta for evolution predictions.