Turn reaction density notebooks into scripts
... to make them more flexible. For example, we can now directly compare
performance on both datasets with the same prediction pipeline. In future,
this should also make it easy to compare different feature sets. We have
checked that the piepline still yields the same results (if only enabling
one of the datasets -- as we only set a seed once, results on the 2nd
dataset can change slightly depending on whether the model is deterministic
or not).