Create first version of experiment pipeline
... for experiments with synthetic constraints. Use one demo dataset and
all the constraint types as defined in the current version of the paper
(apart from mixed constraints). Additionally, we allow more flexibility in
the constructors of the constraint generators (variable-length named
arguments) and fix a small bug in constraint evalution (the result list
was empty because we forgot to append individual results).