Parallelize main experiment pipeline
Use Python's "multiprocessing" library. As this library might not be able
to serialize all kinds of objects for the spawned processes (e.g., our
constraint generators with references to Z3 and underlying C objects), we
push constraint generation into a function which is executed independently
for each generator and dataset. Also, note that we need a main method to
run the code on Windows (as the spawned processes imports the file
containing the function to be parallelized and we don't want to have a
recursion).
Our code is already very general regarding the datasets now (supports a
list of them), so adding new datasets might also be easier in future.
The progress bar also still works (though some adaptations were necessary).
The "results" object is the same as for the old pipeline (apart from the
new column for the dataset).