Add classifiers to prediction pipeline
As for the feature qualities, classifier are represented by a dict
containing the necessary functions (in constrast to the qualities, here we
also have additional arguments). Classifiers can be selected in the main
pipeline. By default, all are used. The number of splits in cross-
validation also is an command line argument.
To prevent leakage, we cannot use pre-computed feature qualities any more,
but re-compute the qualities on each train split (which takes a low amount
of time). Also, we harmonize some naming of the experimental dimensions
(dataset_name, quality_name, constraint_name).

The extended pipeline still returns the same result as the previous version
for our boston demo dataset if the number of splits is set to 0, i.e.,
disabling splitting (and ignoring column renaming, the new UNCONSTRAINED
type as well as additional columns for the prediction score in the
resulting data frame).