Fix objective function bug if quality == 1
If a quality was exactly 1, multiplication of that quality with a z3 bool
variable resulted in another boolean (instead of arithmetic) expression,
which could not be summed for the target function. Thus, we use the more
general arithmetic if-then-else statement, to which all the other bool/
arithmetic products were transformed anyway.