Add MS constraint evaluator
Instead of synthetic constraint generation, we use actual constraints for
the materials science use case. Thus, we need no repeated generation to
fight randomness, but exactly evaluate the desired constraints once. Apart
from that, the solution is object-oriented like the generation approach
and the result is similar, just that it is a dict instead of a DataFrame.
Sub-classes need to specify which constraints to add. We add two
sub-classes, one for the case without constraint and one representing the
only constraint we have from the materials scientists. The pipeline needs
yet to be extended.