Heavily refactor MS pipeline
Integrate constraint evaluator and re-design the pipeline such that it is
similar to the pipeline for experiments with synthetic constraints. In
particular, access datasets, prediction models and quality measures in the
same way. Access constraint evaluator in a similar way as the constraint
generators in the other pipeline. Parallelize the pipeline. Provide
command line arguments. There are still some differences, e.g., each
constraint set is only evaluated once, the train-test split is data-driven
instead of random, we only parallelize over datasets instead of constraint
types etc.
The corresponding pipeline evaluation snippets will be updated soon.