Normalize objective values for evaluation
Objective values differ between datasets (and quality metrics, of which we
only have one), as features might be more or less strongly correlated to
their respective target and even the number of features varies. Thus,
comparisons of absolute objective values are bound to fail. Instead, we
normalize objective values to [0,1] before evaluating results.
In theory, for our simple optimization problem, we could also compute the
max objective by hand in file "combi_solving", but for more complex
problems that would not be possible anyway the solution used now already is
the best way to go.