Make global-AT-MOST stop at k-1 instead of k
Of couse, one can also use the constraint of selecting at most k out of k
features, but this amounts to the unconstrained case and is already covered
by a dedicated constraint "generator". Thus, we decrease the upper
iteration bound by 1. This is also consistent to the primitive AT-MOST
generator, which, if it does not get a cardinality provided, uses at most
num_variables-1 as cardinality.