Remove misleading parameters in MIXED generator
9ae19732ff tried to adapt the parametrization of the sub-generators used in
MixedGenerator. Actually, setting these parameters only matters if the
evaluation function of the sub-generators is used, because that function
decides on the number of constraints and variables in each iteration.
However, MixedGenerator runs its own evaluation function and just calls
the generation functions of the sub-generators. The generation functions
already get a pre-defined group of features and cannot change the
cardinality of that group (apart from XOR, which is only defined for two
variables and therefore always picks the first two variables, no matter how
many are passed).
This behaviour does not really hurt evaluation, we just need to be aware of
it. Thus, we remove the parameters which have no impact.
If we wanted to enforce the strict use of single-IFF and single-NAND in
MIXED, we would either need to define dedicated generators for that or add
a cardinality check in generate() / move constraint-variable selection into
generate().