Fix name clash in arithmetic expressions
Before we made all fields public in cdf9bbb7b0, we had a method value()
in ArithmeticExpression, which usually dispatched to subclasses (e.g., for
a comparison, both sides need to be determined before doing the operation
itself), but returned the attribute __value for the class NumericConstant.
However, after making all fields public, the new attribute name "value"
clashed with the method name, resulting in a TypeError "'int' object is not
callable". Thus, we rename the method.