Fix feature definition for relative predictions
When introducing aggregate functions for the MS datasets in 610af4141e, we
misspelled a string to exclude certain columns from the feature set. This
had the consequence that the target variable had appeared twice in the
pre-processed dataset. In theory, the target was also part of the feature
set due to that mistake, but a later statement in the function (after
adding the aggregates) prevented that bug (making the remaining bug harder
to find).