Make MS constraints more robust
Depending on the features available after pre-processing, not all
constraints might apply to a certain dataset. Most of our constraints use
regular expressions to define groups of features. Constraints build on
these groups. The operators AND, IFF and OR are able to deal with empty
groups by default (in Z3 and our expressions module): AND/IFF over an
emtpty set evaluate to True, OR over an empty set evalutes to False. With
this behavior, some constraints might just evalute to true (are satisfied
trivially) if there are no suitable features to be used. There are no
constraints evaluating to false if features are missing. However, there
were potential issues with AT-MOST, which is undefined for Z3 (while it
evaluates to true in our module combi_expressions if applied to an empty
set of variables). Thus, we make sure AT-MOST only gets fed in a non-empty
list of variables. If the list of variables is empty, we simply not create
a constraint (which has the same effect as a satisfied constraint, apart
from the constraint count).