Add combined evaluator
Up to now, we only had classes for individual constraints in materials
science (though some of the classes added multiple constraints of the same
type, e.g., by looping over physical quantities). It might also be
interesting to arbitrarily combine constraints, without needing to create
a new class for each combination. Thus, we add a new container evaluator
which instantiates multiple sub-evaluators and simply adds constraints from
all of them. The sub-evaluators can be specified during the constructor
call. For serialization purposes (internally, each evaluator contains a
Problem instance and thereby also an Z3 optimizer and Z3 constraints) in
the parallelized experimental pipeline, sub-evaluators are passed as type,
argument pairs instead of instances. This is the same behavior we also have
for the existing evaluators in the MS pipeline and generators in the
synthetic pipeline - the only difference is that for them the name is
passed as string and not as type, as we found out about the latter
possibility just a short time ago.