Add new evaluators to MS pipeline
We already defined several constraint evaluators for the materials science
pipeline a while ago (96ac1dd4bb). However, we didn't include them in the
pipeline until now. We directly combine all evaluators with a global
cardinality constraint.
As another (rather small) change, we include the selected features in the
result, to allow for qualitative evaluation.