Add filter for low-quality features
Some features in the dataset might have very low quality, e.g., a
correlation close to 0. If there are many such features, they slow down
the optimizer, but have a low chance of being selected and even if they
are selected, one cannot really benefit from that, as the correlation might
be spurious. Thus, we add a function to filter out such low-quality
features and integrate that function into the ML pipeline (considering the
train-test split).