Express quality filter with constraints
The filter for low-quality features (introduced in 9b2580e42) was
implemented as a separate function called from the pipeline. While this
worked fine and only marginally extended the pipeline code, it was
somewhat outside the constraint framework. Actually, filtering out low-
quality features is possible within the framework, though rather boring
(if quality under threshold, do not select feature). Nevertheless, for
consistency reasons, we include these constraints in the framework.