Make correlation removal an evaluator
We already had removal of correlated features as a function which goes over
the correlation matrix and removes one of two correlated features.
However, we can also push correlation removal into the CFFS optimization
problem. This might have the additional benefit that the less valuable
feature is removed instead of a random one.
The only tricky thing is that this evaluator depends on feature values and
not just on feature names (only the feature names and their qualities are
usually accessible from the optimization problem). Thus, we cannot define
the new evaluator together with all other evaluators at beginning, but need
to push it deeper into the pipeline. We can simply add the new evaluator to
all base evaluators, as we already need to for global cardinality
constraint and the low-quality-removal constraint.