Introduce Boolean values as expressions
Though we had the possibility to create numeric constants in our constraint
framework, we did not provide the option to create a Boolean constant. For
completeness (we don't actually use it right now), we add this option. We
also do some other refactorings, which should not change the behavior of
the program.