Add a mixed evaluator to MS pipeline
For synthetic constraints, we already habe a generator which combines
multiple basic constraint types. In the materials science, we only had
combinations with global cardinality, low-quality threshold and correlation
threshold for all basic evaluators, but no real combination of the basic
evaluators. We add such a combination now, again making use of our
CombinedEvaluator. For the mixed constraints, the new evaluator also gets
nested in the existing CombinedEvaluator that adds cardinality constraints
etc. This demonstrates the flexibility of our framework.