Fix access to z3_expr in combi_solving
When renaming the attribute "z3" of combi_solving.BooleanExpression to
"z3_expr" in 9776aaca12, we adapted setter access, but not getter access.
We fix this now and additionally make sure that getter access goes through
the corresponding method instead of accessing the attribute directly
(though the getter does not provide additional functionality at the
moment).