Enable counting of constrained variables
The number of variables involved in constraints might be an interesting
evaluation metric. However, our constraints are composite expressions,
which were primarility intended for evaluating their value. Thus, we add
a method which allows to traverse the expression tree and retrieves the
variables. We call the method in the constraint evaluation.