Compute relative feature-count metrics for evaluation
Taking the new metrics introduced in 1e4a7611c3, we add a routine to
normalize them with the number of variables. We also move the comutation of
the relative number of selected features to the new routine (before, it was
kind of redundant in the class "Problem", as the absolute number of
selected features was stored as well).