Track variable-count metrics for NoConstraintGenerator
Since commit 6a513cc679, evaluation metrics normalized with the number
of variables are not computed in the constraint generators themselves, but
only afterwards. However, the number of variables and of constrained
variables has to be tracked in the generator. This worked for nearly all
generators of synthetic constraints, as they use the same evaluation
routine from a super-class. However, for the unconstrained case, the
evaluation routine is overridden, so only the optimizer is run without
constraints and without repeated generation. Most of the evaluation metrics
are set manually. We simply forgot to set the variable-related evaluation
metrics. This resulted in some NAs in the results file.
The materials-science study is not affected, as the unconstrained evaluator
there uses the standard evaluation routine from its super-class.