Fix feature order in correlation-removal evaluator
When introducing the constraint evaluator to remove inter-feature
correlations (commit aaa94a119b), we checked that the correlation matrix
has the same number of features as the optimization problem. However, we
did not make sure that the order of features is the same. This assumption
introduced errors, as the constructor of class `Problem` actually re-orders
features (together with their qualities) to improve computational
efficiency.
As solution, we switch from position-based indexing to name-based indexing
for creating the correlation constraints. This would even work for if names
partially differed (e.g., correlation matrix only for some features), since
only constraints for matching names would be added in that case. However,
to avoid further surprises, we assert that all feature names are the same
in the optimization problem and the correlation matrix.