Review code in package "materials_science"
No bugs detected. Only minor refactoring, which should not affect the
outcome of the experiments. Added some documentation.

Changes:
- Let main scripts print a starting message (not only an end message).
- Add more references to paper in evaluation script.
- Remove some unused code in MS data utility.
- Remove abstract super-class ContainerEvaluator.
- Rename domain-independent MS evaluators to be consistent to the paper.
- Simplify PlasticStrainTensorEvaluator, i.e., choose between features and
not feature groups. This is possible because we already removed history/
neighbor features. Also, this is consistent to the paper.
- Simplify DislocationDensityEvaluator. Before, it formed three groups,
two consisting of one feature and one consisting of multiple features.
Choosing between these groups was not really fair due to the different
sizes. Thus, we flatten the constraint the choose between all features.
Paper needs to be adapted accordingly.