Fix creation of results dir in MS pipeline
In 7fac20c754, we deleted the code which creates a MS results dir if not
existing. We thought it unnecessary in the pipeline() function, as the
directory is not used there. While this is true, the results are eventually
stored, but in the main routine. Thus, we add the dir creation code there.