Move creation of results dir in syn pipeline
While moving the dir creation into the pipeline() routine in 10c2810f60 did
not cause any error yet (unlike for the MS pipeline, where we had to fix a
bug in ba30bf9f11, after we not only moved the code, but later deleted it),
this refactoring change still was not optimal: The results dir is also used
if __main__, and it actually just is optional in pipeline(). Thus, it
should be created in __main__. As a consequence, we move the code back to
__main__. The check for data_dir still remains in pipeline(), as data_dir
is not used in main. The same principle is applied in the MS pipeline.